Combating freefriders and evil Bobs with a KFrag writ authorization.
