+++ nucypher/blockchain/eth/actors.py
@@ -21,8 +21,8 @@ import json
 import maya
 import os
 import time
-from bytestring_splitter import BytestringSplitter
 from constant_sorrow.constants import FULL, WORKER_NOT_RUNNING
+from datetime import datetime
 from decimal import Decimal
 from eth_tester.exceptions import TransactionFailed as TestTransactionFailed
 from eth_typing import ChecksumAddress
@@ -58,8 +58,7 @@ from nucypher.blockchain.eth.constants import (
     NULL_ADDRESS,
     EMERGENCY_MANAGER,
     STANDARD_AGGREGATOR,
-    STANDARD_VOTING,
-    POLICY_ID_LENGTH
+    STANDARD_VOTING
 )
 from nucypher.blockchain.eth.decorators import (
     only_me,
@@ -100,6 +99,7 @@ from nucypher.characters.banners import STAKEHOLDER_BANNER
 from nucypher.characters.control.emitters import StdoutEmitter
 from nucypher.config.constants import DEFAULT_CONFIG_ROOT
 from nucypher.crypto.powers import TransactingPower
+from nucypher.policy.policies import Policy
 from nucypher.types import NuNits, Period
 from nucypher.utilities.logging import Logger
 
@@ -117,7 +117,8 @@ class BaseActor:
                  domain: Optional[str],
                  registry: BaseContractRegistry,
                  transacting_power: Optional[TransactingPower] = None,
-                 checksum_address: Optional[ChecksumAddress] = None):
+                 checksum_address: Optional[ChecksumAddress] = None,
+                 economics: BaseEconomics = None):
 
         if not (bool(checksum_address) ^ bool(transacting_power)):
             error = f'Pass transacting power or checksum address, got {checksum_address} and {transacting_power}.'
@@ -135,6 +136,7 @@ class BaseActor:
             else:
                 self.checksum_address = checksum_address
 
+        self.economics = economics or StandardTokenEconomics()
         self.transacting_power = transacting_power
         self.registry = registry
         self.network = domain
@@ -221,9 +223,8 @@ class ContractAdministrator(BaseActor):
     class UnknownContract(ValueError):
         pass
 
-    def __init__(self, economics: BaseEconomics = None, *args, **kwargs):
+    def __init__(self, *args, **kwargs):
         self.log = Logger("Deployment-Actor")
-        self.economics = economics or StandardTokenEconomics()
         self.deployers = {d.contract_name: d for d in self.all_deployer_classes}
         super().__init__(*args, **kwargs)
 
@@ -1041,8 +1042,6 @@ class Worker(NucypherTokenActor):
     READY_POLL_RATE = 10
     READY_CLI_FEEDBACK_RATE = 60  # provide feedback to CLI every 60s
 
-    policy_id_splitter = BytestringSplitter(POLICY_ID_LENGTH)
-
     class WorkerError(NucypherTokenActor.ActorError):
         pass
 
@@ -1050,11 +1049,6 @@ class Worker(NucypherTokenActor):
         """Raised when the Worker is not bonded to a Staker in the StakingEscrow contract."""
         crash_right_now = True
 
-    class UnpaidPolicy(WorkerError):
-        """Raised when a worker expects policy payment but receives none."""
-
-    class UnknownPolicy(WorkerError):
-        """Raised when a worker cannot find a published policy for a given policy ID"""
 
     def __init__(self,
                  is_me: bool,
@@ -1191,18 +1185,25 @@ class Worker(NucypherTokenActor):
         missing = self.staking_agent.get_missing_commitments(checksum_address=staker_address)
         return missing
 
-    def verify_policy_payment(self, policy_id: bytes) -> bool:
-        arrangements = self.policy_agent.fetch_policy_arrangements(policy_id=policy_id)
+    def verify_policy_payment(self, hrac: bytes) -> None:
+        arrangements = self.policy_agent.fetch_policy_arrangements(policy_id=hrac)
         members = set()
         for arrangement in arrangements:
             members.add(arrangement.node)
             if self.checksum_address == arrangement.node:
-                return True
+                return
         else:
-            policy_id_hex = policy_id.hex()
             if not members:
-                raise self.UnknownPolicy(f'{policy_id_hex} is not a published policy.')
-            raise self.UnpaidPolicy(f"{policy_id_hex} is unpaid.")
+                raise Policy.Unknown(f'{hrac.hex()} is not a published policy.')
+            raise Policy.Unpaid(f"{hrac.hex()} is unpaid.")
+
+    def verify_active_policy(self, hrac: bytes) -> None:
+        policy = self.policy_agent.fetch_policy(policy_id=hrac)
+        if policy.disabled:
+            raise Policy.Inactive(f'{hrac.hex()} is a disabled policy.')
+        expired = datetime.utcnow() >= datetime.utcfromtimestamp(policy.end_timestamp)
+        if expired:
+            raise Policy.Expired(f'{hrac.hex()} is an expired policy.')
 
 
 class BlockchainPolicyAuthor(NucypherTokenActor):
