+++ nucypher/blockchain/eth/constants.py
@@ -56,8 +56,7 @@ MAX_UINT16 = 65535
 NULL_ADDRESS = '0x' + '0' * 40
 
 # NuCypher
-
-POLICY_ID_LENGTH = 16
+POLICY_ID_LENGTH = 16  # same as the HRAC size
 
 # Aragon DAO contract names
