+++ tests/acceptance/characters/control/test_web_control_blockchain.py
@@ -26,7 +26,7 @@ from click.testing import CliRunner
 import nucypher
 from nucypher.crypto.kits import UmbralMessageKit
 from nucypher.crypto.powers import DecryptingPower
-from nucypher.policy.collections import TreasureMap, SignedTreasureMap
+from nucypher.policy.maps import TreasureMap, SignedTreasureMap
 
 click_runner = CliRunner()
