+++ tests/integration/network/test_network_upgrade.py
@@ -23,20 +23,6 @@ from cryptography.hazmat.primitives import serialization
 from twisted.internet import threads
 
 from nucypher.characters.lawful import Ursula
-from nucypher.datastore.models import PolicyArrangement
-from tests.utils.ursula import make_federated_ursulas
-
-
-def test_alice_enacts_policies_in_policy_group_via_rest(enacted_federated_policy, federated_ursulas):
-    """
-    Now that Alice has made a PolicyGroup, she can enact its policies, using Ursula's Public Key to encrypt each offer
-    and transmitting them via REST.
-    """
-    for ursula in federated_ursulas:
-        arrangement_id = enacted_federated_policy.treasure_map.destinations[ursula.checksum_address]
-        with ursula.datastore.describe(PolicyArrangement, arrangement_id.hex()) as policy_arrangement:
-            the_kfrag = policy_arrangement.kfrag
-        assert bool(the_kfrag)  # TODO: This can be a more poignant assertion.
 
 
 @pytest_twisted.inlineCallbacks
