+++ tests/utils/ursula.py
@@ -14,7 +14,7 @@ GNU Affero General Public License for more details.
 You should have received a copy of the GNU Affero General Public License
 along with nucypher.  If not, see <https://www.gnu.org/licenses/>.
 """
-
+import os
 
 import contextlib
 import socket
@@ -31,7 +31,8 @@ from nucypher.characters.lawful import Bob
 from nucypher.characters.lawful import Ursula
 from nucypher.config.characters import UrsulaConfiguration
 from nucypher.crypto.utils import canonical_address_from_umbral_key
-from nucypher.policy.collections import WorkOrder, IndisputableEvidence
+from nucypher.policy.disputes import IndisputableEvidence
+from nucypher.policy.orders import WorkOrder
 from tests.constants import NUMBER_OF_URSULAS_IN_DEVELOPMENT_NETWORK
 from tests.mock.datastore import MOCK_DB
 
@@ -182,16 +183,13 @@ def _mock_ursula_reencrypts(ursula, corrupt_cfrag: bool = False):
                                  sign_receiving_key=False)
     capsule.set_correctness_keys(delegating_privkey.get_pubkey(), pub_key_bob, signing_pubkey)
 
+    # TODO: Is this needed?
+    # alice_address = canonical_address_from_umbral_key(signing_pubkey)
     ursula_pubkey = ursula.stamp.as_umbral_pubkey()
-
-    alice_address = canonical_address_from_umbral_key(signing_pubkey)
-    blockhash = bytes(32)
-
     specification = b''.join((bytes(capsule),
                               bytes(ursula_pubkey),
-                              bytes(ursula.decentralized_identity_evidence),
-                              alice_address,
-                              blockhash))
+                              bytes(ursula.decentralized_identity_evidence)
+                              ))
 
     bobs_signer = Signer(priv_key_bob)
     task_signature = bytes(bobs_signer(specification))
@@ -207,15 +205,15 @@ def _mock_ursula_reencrypts(ursula, corrupt_cfrag: bool = False):
 
     bob = Bob.from_public_keys(verifying_key=pub_key_bob)
     task = WorkOrder.PRETask(capsule, task_signature, cfrag, cfrag_signature)
-    hrac = bob.construct_policy_hrac(verifying_key=signing_pubkey, label=b'this is the label')
+    hrac = bob.construct_policy_hrac(relayer_verifying_key=signing_pubkey, label=b'this is the label')
     work_order = WorkOrder(bob=bob,
                            hrac=hrac,
-                           encrypted_kfrag=None,
+                           encrypted_kfrag=os.urandom(606),
                            tasks={capsule: task},
                            receipt_signature=None,
                            ursula=ursula,
-                           authorizer_verifying_key=signing_pubkey.to_bytes(),
-                           publisher_verifying_key=signing_pubkey.to_bytes())
+                           alice_verifying_key=signing_pubkey.to_bytes(),
+                           relayer_verifying_key=signing_pubkey.to_bytes())
 
     evidence = IndisputableEvidence(task, work_order)
     return evidence
